+++ statik/deploy/sftp.py
@@ -1,0 +1,10 @@
+"""
+Deployment support for SFTP.
+"""
+
+from .base import DeploymentMethod
+
+
+class SFTP(DeploymentMethod):
+    pass
+